Changed: WorkProduct to WorkProductService (#1280)
Because of changes in the past WorkProduct and the associated subclasses do not
represent the actual work product anymore, instead they are services to work with
the work product. To clarify this, this commit renames all the WorkProduct classes
to WorkProductService.

This commit also changes the usage of these classes to use injection instead of newing
the class up each time it is used.