Set User object as CurrentUser on every request. (#1635)
* Set `User` object as `CurrentUser` on every request.

We previously made heavy use of the `ProxyUser` class, which prevented
heavy-weight objects from being stored in the session and also prevented
the loading of a "real" `User` object from the database if all that was
required was the userid. However, after some investigation, it turns out
that we almost always load a real `User` object anyway. Additionally, we
no longer store anything in the `HttpSession`, so there's really no need
for the ProxyUser.

This commit removes the `ProxyUser` and, despite being a large commit,
simplifies a lot of things in the process.

* Unintentionally changed method visibility to private

* Remove TestUser in favor of using existing InMemoryUser

* Clean things up with a single arg constructor